Logic for checking if index for next pointer created

+++ distrib/os/shell.js
@@ -366,19 +366,34 @@ var TSOS;
         };
         Shell.prototype.shellCreate = function (args) {
             if (args.length > 0) {
-                _DeviceDiskDriver.createFile(args.toString());
+                if (_FORMATTED) {
+                    if (_DeviceDiskDriver.createFile(args.toString())) {
+                        _StdOut.putText("Created file " + args.toString());
+                    }
+                    else {
+                        _StdOut.putText("No, " + args.toString() + " Already exists");
+                    }
+                }
+                else {
+                    _StdOut.putText("Please format hard drive.");
+                }
             }
             else {
-                _StdOut.putText("Disk Already Formatted");
+                _StdOut.putText("create <fileName>");
             }
         };
         Shell.prototype.shellDelete = function (args) {
             if (args.length > 0) {
-                if (_DeviceDiskDriver.deleteFile(args.toString())) {
-                    _StdOut.putText("Deleted " + args.toString());
+                if (_FORMATTED) {
+                    if (_DeviceDiskDriver.deleteFile(args.toString())) {
+                        _StdOut.putText("Deleted " + args.toString());
+                    }
+                    else {
+                        _StdOut.putText("Could not delete " + args.toString());
+                    }
                 }
                 else {
-                    "Could not delete " + args.toString();
+                    _StdOut.putText("Please format hard drive.");
                 }
             }
             else {
